code/requirements.txt
cassandra-driver-3.25.0